requirements.txt
transformers==4.41.2
torch==2.3.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.0
pyyaml==6.0.1
aiohttp==3.10.11
shodan==1.31.0
requests==2.32.3
yara-python==4.5.1
python-magic==0.4.27
pyOpenSSL==24.1.0
datasets==2.19.1
pytest==8.2.1
minio==7.2.7
nmap-python==0.6.1
logging==0.5.1.2
pydantic==2.7.1
fastapi==0.110.2
uvicorn==0.29.0